+++ gruntfile.js
@@ -1,15 +1,38 @@
 // jshint ignore: start
 module.exports = function (grunt) {
 
-    // Project configuration.
     grunt.initConfig({
         pkg: grunt.file.readJSON('package.json'),
+        imports: 'jQuery, window, document, Object, String, Array, Math, Node, Range, DocumentFragment, RegExp',
         concat: {
-            options: {
-                process: true,
+            shim: {
+                options: {
+                    banner: 'new (function () {\n\n',
+                    footer: '\n})',
+                    process: function(src, filepath) {
+                        return '// source: ' + filepath + '\n' + src;
+                    }
+                },
+                src: ['src/shim/*', 'node_modules/setimmediate/setimmediate.js'],
+                dest: 'build/shim.js'
             },
-            build: {
-                src: ['src/license.js', 'src/{core,presets,ui}.js', 'src/canvas.js', 'src/{themes,extensions,presets}/*', 'node_modules/setimmediate/setimmediate.js'],
+            lib: {
+                options: {
+                    banner: '(function (<%= imports %>, shim) {\n\'use strict\';\n\n',
+                    footer: '\n}(<%= imports %>, <%= grunt.file.read("build/shim.js") %>));\n',
+                    process: function(src, filepath) {
+                      src = src.replace(/(\n[ \t]*)(?:'use strict'|"use strict");?\s*/g, '$1');
+                      return '// source: ' + filepath + '\n' + src;
+                    }
+                },
+                src: ['src/{core,presets,ui}.js', 'src/canvas.js', 'src/{themes,extensions,presets}/*'],
+                dest: 'build/lib.js'
+            },
+            dist: {
+                options: {
+                    process: true
+                },
+                src: ['src/license.js', 'build/lib.js'],
                 dest: 'dist/jquery.typer.js'
             }
         },
@@ -22,14 +45,14 @@ module.exports = function (grunt) {
                 src: 'dist/jquery.typer.js',
                 dest: 'dist/jquery.typer.min.js'
             }
-        }
+        },
+        clean: ['build/']
     });
 
-    // Load the plugin that provides the "uglify" task.
     grunt.loadNpmTasks('grunt-contrib-concat');
     grunt.loadNpmTasks('grunt-contrib-uglify');
+    grunt.loadNpmTasks('grunt-contrib-clean');
 
-    // Default task(s).
-    grunt.registerTask('default', ['concat', 'uglify']);
+    grunt.registerTask('default', ['concat:shim', 'concat:lib', 'concat:dist', 'uglify', 'clean']);
 
 };