+++ src/extensions/link.js
@@ -184,4 +184,4 @@
         }
     });
 
-}(jQuery, window.Typer));
+}(jQuery, Typer));