+++ src/extensions/validation.js
@@ -35,4 +35,4 @@
         }
     };
 
-} (jQuery, window.Typer));
+}(jQuery, Typer));