+++ src/presets/keyword.js
@@ -272,4 +272,4 @@
         }
     };
 
-}(jQuery, window.Typer));
+}(jQuery, Typer));