+++ src/presets/textbox.js
@@ -29,4 +29,4 @@
         }
     };
 
-} (jQuery, window.Typer));
+}(jQuery, Typer));