+++ src/shim/Set.js
@@ -1,0 +1,34 @@
+this.Set = window.Set || (function () {
+    function Set() {
+        this.items = [];
+    }
+    Set.prototype = {
+        get size() {
+            return this.items.length;
+        },
+        add: function (v) {
+            var items = this.items;
+            if (items.indexOf(v) < 0) {
+                items.push(v);
+            }
+            return this;
+        },
+        delete: function (v) {
+            var index = this.items.indexOf(v);
+            if (index >= 0) {
+                this.items.splice(index, 1);
+            }
+            return index >= 0;
+        },
+        forEach: function (callback, thisArg) {
+            var self = this;
+            self.items.forEach(function (v) {
+                callback.call(thisArg, v, v, self);
+            });
+        },
+        clear: function () {
+            this.items.splice(0);
+        }
+    };
+    return Set;
+}());